implementations/feature4_restaurant_notifications/requirements.txt
Flask==2.0.1
redis==4.5.1
psycopg2-binary==2.9.5
python-dotenv==0.19.0